erro 404 RESOLVIDO

+++ forum/src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -17,6 +17,7 @@ import javax.transaction.Transactional;
 import javax.validation.Valid;
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 //TODO METODO TEM O RESPONSE BODY
 @RestController
@@ -60,28 +61,48 @@ public class TopicosController {
     }
 
     @GetMapping("/{id}")
-    public DetalhesdoTopicoDto detalhar(@PathVariable  Long id)
+    public ResponseEntity<DetalhesdoTopicoDto> detalhar(@PathVariable  Long id)
     {
-        Topico topico=topicoRepository.getOne(id);
-        return new DetalhesdoTopicoDto(topico);
+        //PÓDE SER QUE TENHA UM REGISTRO OU NÃO
+        Optional<Topico> topico=topicoRepository.findById(id);
+        if(topico.isPresent())
+        {
+            return ResponseEntity.ok(new DetalhesdoTopicoDto(topico.get()));
+        }
+
+        return  ResponseEntity.notFound().build();
     }
 
     @PutMapping("/{id}")
     //COMITAR A TRANSAÇÃO
     @Transactional
     public ResponseEntity<TopicoDto>  atualizar(@PathVariable  Long id,@RequestBody @Valid Atualizacaoform form){
-        Topico topico=form.atualizar(id,topicoRepository);
+        Optional<Topico> optional=topicoRepository.findById(id);
+        if(optional.isPresent())
+        {
+            Topico topico=form.atualizar(id,topicoRepository);
+            return ResponseEntity.ok(new TopicoDto(topico));
+        }
 
-        return ResponseEntity.ok(new TopicoDto(topico));
+
+        return ResponseEntity.notFound().build();
+
     }
 
 
     @DeleteMapping("/{id}")
     @Transactional
     public ResponseEntity remover(@PathVariable  Long id){
-        topicoRepository.deleteById(id);
 
-        return ResponseEntity.ok().build();
+        Optional<Topico> optional=topicoRepository.findById(id);
+        if(optional.isPresent())
+        {
+            topicoRepository.deleteById(id);
+
+            return ResponseEntity.ok().build();
+        }
+
+        return ResponseEntity.notFound().build();
     }
 
 }
